move config to @io-ipatente/core
