+++ apps/vehicles/src/app/api/auth/mocks/oidc/token/route.ts
@@ -1,5 +1,5 @@
 import { getMockForbiddenResponse } from "@/app/api/auth/mocks/mocks-util";
-import { getConfiguration } from "@/config";
+import { getConfiguration } from "@io-ipatente/core";
 import jwt from "jsonwebtoken";
 import { NextResponse } from "next/server";
