+++ apps/vehicles/src/lib/bff/client.ts
@@ -1,5 +1,5 @@
-import { getConfiguration } from "@/config";
 import { createApiClient } from "@/generated/ext-openapi";
+import { getConfiguration } from "@io-ipatente/core";
 import { ApiOf, ZodiosInstance } from "@zodios/core";
 
 const { EXT_API_BASE_PATH, EXT_API_BASE_URL, EXT_API_MOCKING } =
