+++ apps/vehicles/src/lib/bff/with-voucher-handler.ts
@@ -1,4 +1,4 @@
-import { getConfiguration } from "@/config";
+import { getConfiguration } from "@io-ipatente/core";
 import {
   Voucher,
   generateClientAssertion,
