+++ packages/core/lib/config/index.ts
@@ -1,2 +1,1 @@
-export * from "./bff";
 export * from "./config";
